# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Security Guard ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payroll:
    get:
      operationId: getPayrolls
      tags: [payroll]
      summary: List payroll records
      parameters:
        - name: guardId
          in: query
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payroll records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payroll"
    post:
      operationId: createPayroll
      tags: [payroll]
      summary: Create payroll record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
  /payroll/{id}:
    get:
      operationId: getPayroll
      tags: [payroll]
      summary: Get payroll record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payroll record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
    patch:
      operationId: updatePayroll
      tags: [payroll]
      summary: Update payroll record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ─── EXPENSES ──────────────────────────────────────────────────────────────
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: submittedBy
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # ─── EQUIPMENT ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Payroll:
      type: object
      required: [id, guardId, period, basicSalary, netSalary, status, createdAt]
      properties:
        id:
          type: integer
        guardId:
          type: integer
        guardName:
          type: ["string", "null"]
        period:
          type: string
        basicSalary:
          type: number
        allowances:
          type: number
        deductions:
          type: number
        netSalary:
          type: number
        status:
          type: string
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
    PayrollInput:
      type: object
      required: [guardId, period, basicSalary]
      properties:
        guardId:
          type: integer
        period:
          type: string
        basicSalary:
          type: number
        allowances:
          type: number
        deductions:
          type: number
        status:
          type: string
    PayrollUpdate:
      type: object
      properties:
        basicSalary:
          type: number
        allowances:
          type: number
        deductions:
          type: number
        status:
          type: string
        paidAt:
          type: string
...
    Invoice:
      type: object
      required: [id, clientId, invoiceNumber, amount, status, dueDate, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        period:
          type: ["string", "null"]
        amount:
          type: number
        tax:
          type: number
        totalAmount:
          type: number
        status:
          type: string
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceInput:
      type: object
      required: [clientId, invoiceNumber, amount, dueDate]
      properties:
        clientId:
          type: integer
        invoiceNumber:
          type: string
        period:
          type: string
        amount:
          type: number
        tax:
          type: number
        dueDate:
          type: string
        notes:
          type: string
        status:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        amount:
          type: number
        tax:
          type: number
        status:
          type: string
        dueDate:
          type: string
        paidAt:
          type: string
        notes:
          type: string

    # ── Expense ───────────────────────────────────────────────────────────────
    Expense:
      type: object
      required: [id, category, amount, date, status, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        submittedBy:
          type: ["integer", "null"]
        submittedByName:
          type: ["string", "null"]
        approvedBy:
          type: ["integer", "null"]
        receiptUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [category, amount, date]
      properties:
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        submittedBy:
          type: integer
        receiptUrl:
          type: string
        status:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        status:
          type: string
        approvedBy:
          type: integer

    # ── Equipment ─────────────────────────────────────────────────────────────
...