# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Security Guard ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  category?: string;
  amount?: number;
  description?: string;
  status?: string;
  approvedBy?: number;
}